Some modernisations
* Switch to edge for default branch

Already made the change on GitHub, but this will update the badge for
the README as well.

* use my new style of gem rakefile

* stop printing rspec in a super verbose way

* Use the fun new gems.rb file over Gemfile

* update rubocop and refactor a bit

this makes one method private which was supposed to be private all along
(::uri). That's technically a breaking change. But I'm not sure I care!

* Remove rdio -- service was acquired and shut down!

* Fix the instagram endpoint

I guess it changed again?

* Fix a youtube spec after a video was deleted

damn you Pete!

* Fix another broken spec caused by deleted thing

Damn you dom! you deleted the very first vine! haha

* Test fewer rubies -- relax a little

* make sure the gems are installed